Calculate annual anomaly

Calculate annual anomaly

calc_annual_anomaly(
  x,
  year,
  period = NULL,
  max_missing = 0.2,
  max_consecutive = NULL
)

Arguments

x

metric. A vector

year

the year in which the metric was observed. A vector of the same length as x

period

the reference period of the anomaly. A vector of length 2 or NULL. If a vector, the first element indicates the beginning of the reference period and the second the end of the reference period. If NULL (the default) the reference period goes from the first to the last year.

max_missing

Numeric. Maximum number or proportion of missing values in x. If any number of missing values is allowed set to 1 (100%) or NULL.

max_consecutive

Numeric. Maximum number or proportion of consecutive missing values in x. If any number of missing values is allowed set to 1 (100%) or NULL.

Value

a vector with the anomaly of x

Examples

temperature_frame <- tibble::tibble(temperature = runif(100, 10, 20),
                                    year = 1901:2000)
reference_period <- c(1961, 1990)
temperature_frame %>%
dplyr::mutate(anomaly = calc_annual_anomaly(temperature,
                                           year,
                                           reference_period))
#> # A tibble: 100 × 3
#>    temperature  year anomaly
#>          <dbl> <int>   <dbl>
#>  1        10.1  1901 -4.57  
#>  2        14.7  1902  0.0181
#>  3        15.0  1903  0.332 
#>  4        12.9  1904 -1.75  
#>  5        17.3  1905  2.68  
#>  6        17.7  1906  3.08  
#>  7        18.7  1907  4.10  
#>  8        11.7  1908 -2.90  
#>  9        10.3  1909 -4.30  
#> 10        13.2  1910 -1.44  
#> # ℹ 90 more rows