Calculate annual anomaly

Calculate annual anomaly

calc_annual_anomaly(
  x,
  year,
  period = NULL,
  max_missing = 0.2,
  max_consecutive = NULL
)

Arguments

x
metric. A vector

year
the year in which the metric was observed. A vector of the same length as x

period
the reference period of the anomaly. A vector of length 2 or NULL. If a vector, the first element indicates the beginning of the reference period and the second the end of the reference period. If NULL (the default) the reference period goes from the first to the last year.

max_missing
Numeric. Maximum number or proportion of missing values in x. If any number of missing values is allowed set to 1 (100%) or NULL.

max_consecutive
Numeric. Maximum number or proportion of consecutive missing values in x. If any number of missing values is allowed set to 1 (100%) or NULL.

Value

a vector with the anomaly of x

Examples

temperature_frame <- tibble::tibble(temperature = runif(100, 10, 20),
                                    year = 1901:2000)
reference_period <- c(1961, 1990)
temperature_frame %>%
dplyr::mutate(anomaly = calc_annual_anomaly(temperature,
                                           year,
                                           reference_period))

#> # A tibble: 100 × 3
#>    temperature  year anomaly
#>          <dbl> <int>   <dbl>
#>  1        10.8  1901  -3.67 
#>  2        18.3  1902   3.87 
#>  3        16.0  1903   1.53 
#>  4        11.6  1904  -2.91 
#>  5        10.1  1905  -4.40 
#>  6        14.7  1906   0.187
#>  7        15.0  1907   0.501
#>  8        12.9  1908  -1.58 
#>  9        17.3  1909   2.85 
#> 10        17.7  1910   3.25 
#> # … with 90 more rows